# schemas/web-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JeRyu Web Forge API",
    "description": "BFF for the JeRyu Web Forge. Versioned under /api/v1/. See WEB_WORK_CLAUDE.md §35.7 for the canonical route map.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0-alpha"
  },
  "paths": {
    "/api/v1/actions/execute": {
      "post": {
        "tags": [
          "actions"
        ],
        "operationId": "execute_action",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Execution result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing Idempotency-Key"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Unknown action"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
    "/api/v1/actions/preview": {
      "post": {
        "tags": [
          "actions"
        ],
        "operationId": "preview_action",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Action preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "404": {
            "description": "Unknown action"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
...
    "/api/v1/repos/{repo_id}/agents/evidence": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "list_evidence",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent evidence list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentEvidenceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/agents/sessions": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "list_sessions",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent sessions list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSessionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/v1/repos/{repo_id}/issues": {
      "get": {
        "tags": [
          "issues"
        ],
        "operationId": "list_issues",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Not implemented in v1 (see v1.5)"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "issues"
        ],
        "operationId": "create_issue",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Not implemented in v1 (see v1.5)"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/issues/{iid}": {
      "get": {
        "tags": [
          "issues"
        ],
        "operationId": "get_issue",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iid",
            "in": "path",
            "description": "Issue internal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Not implemented in v1 (see v1.5)"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "issues"
        ],
        "operationId": "patch_issue",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iid",
            "in": "path",
            "description": "Issue internal ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Not implemented in v1 (see v1.5)"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentEvidenceItem": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "summary"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          }
        }
      },
      "AgentEvidenceResponse": {
        "type": "object",
        "required": [
          "repo_id",
          "evidence"
        ],
        "properties": {
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentEvidenceItem"
            }
          },
          "repo_id": {
            "type": "string"
          }
        }
      },
...
      "AgentSessionsResponse": {
        "type": "object",
        "required": [
          "repo_id",
          "sessions"
        ],
        "properties": {
          "repo_id": {
            "type": "string"
          },
          "sessions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "ExecuteRequest": {
        "type": "object",
        "required": [
          "action_id"
        ],
        "properties": {
          "action_id": {
            "type": "string"
          },
          "params": {
            "type": "object"
          }
        }
      },
      "ExecuteResponse": {
        "type": "object",
        "required": [
          "action_id",
          "result",
          "event_seq"
        ],
        "properties": {
          "action_id": {
            "type": "string"
          },
          "event_seq": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "result": {
            "type": "object"
          }
        }
      },
...
      "PreviewRequest": {
        "type": "object",
        "required": [
          "action_id"
        ],
        "properties": {
          "action_id": {
            "type": "string"
          },
          "params": {
            "type": "object"
          }
        }
      },
      "PreviewResponse": {
        "type": "object",
        "required": [
          "action_id",
          "preview",
          "event_seq"
        ],
        "properties": {
          "action_id": {
            "type": "string"
          },
          "event_seq": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "preview": {
            "type": "object"
          }
        }
      },
...
    },
    "securitySchemes": {
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jeryu_session"
      }
    }
  },
...
}